refactor(ProductsController): 컨트롤러 분리
