+++ src/main/java/cart/controller/ViewController.java
@@ -1,0 +1,38 @@
+package cart.controller;
+
+import cart.dto.ProductResponse;
+import cart.service.ProductService;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Controller
+public class ViewController {
+
+    private final ProductService productService;
+
+    public ViewController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping("/")
+    public String index(Model model) {
+        List<ProductResponse> products = productService.findProducts().stream()
+                .map(ProductResponse::from)
+                .collect(Collectors.toList());
+        model.addAttribute("products", products);
+        return "index";
+    }
+
+    @GetMapping("/admin")
+    public String admin(Model model) {
+        List<ProductResponse> products = productService.findProducts().stream()
+                .map(ProductResponse::from)
+                .collect(Collectors.toList());
+        model.addAttribute("products", products);
+        return "admin";
+    }
+}
